refactor: PR #119 review 收口（Any 替换 / 抽取重复函数 / 模块职责说明 / 非法 turn warning）
- conversation_archive_init.py：``payload: Any`` → ``object``（硬约束修复）。
- _coercion.py：抽取重复定义的 ``_normalize_session_id`` / ``_utc_now_iso``，
  conversation_store.py 与 conversation_session_archive.py 共用同一份实现。
- conversation_store.py：模块 docstring 显式说明同时承载 transcript 数据类
  与 archive store 抽象/实现的理由，避免与 conversation_session_archive
  形成循环导入。
- ConversationHistoryArchive.from_dict：遇到非 dict turn 元素改为 Log.warning
  + skip，与迁移脚本的 warning 策略对齐，便于排查损坏数据。

测试：4819 passed；pyright 零新增报错。
